Add guest-aware snapshot listener
This listener connects to a gRPC server running on the host and waits
for "pre snapshot requests" and "post snapshot requests". On these
requests, the listener will run scripts as defined in
instance_configs.cfg and send back the result to the host.

Note that this adds a a few snapshot related configuration fields to
instance_configs.cfg, so the documentation at:
https://github.com/GoogleCloudPlatform/compute-image-packages/tree/master/packages/python-google-compute-engine
should be updated.