Pull wsfc configuration values into AgentConfig
In some cases, we rely on knowing whether or not certain wsfc
configuration values are specified in the INI config file. To
ensure this logic still works, I've introduced a new computed
field to the AgentConfig: Wsfc.ExplicitlyConfigured. The value
of this field is set after the INI file is parsed (based on the
contents of the INI file).