Fix loading ip_forwarding configuration value
We have a name collision between the IpForwarding section and the
ip_forwarding key in the INI configuration file. We have no easy way
to map from an IpForwarding field in the AgentConfig to both of these
values.

The current implementation will only map the IpForwarding field to
the IpForwarding section in the INI configuration file. This commit
will manually load the values for the ip_forwarding key.

Also added a test to ensure this doesn't break in the future.