Enable creating Google users with GID.
On COS, /home is stateful but /etc is stateless. So on reboots, all the
users are wiped but their home directories persist. A previous commit
was submitted to create a new user with the UID of their home directory
if it already exists. We need to also create the user with the GID of
their home directory, since it is not necessarily the case that the
UID == GID for a given user.