Avoid conflict with automated package updates
The following services on Debian and RedHat platforms (and their
derivatives) implement automated updates of yum/apt packages:

- yum-cron.service (install new packages on RedHat 7 systems)
- dnf-automatic.service (download and install new packages on RedHat 8+)
- dnf-automatic-install.service (install packages previously downloaded
    by dnf-automatic-download.service)
- apt-daily-upgrade.service (actually installs packages downloaded by
  apt-daily.service)
- unattended-upgrades.serivce (security fixes on Debian)
- google-osconfig-agent.service (Google Cloud tool for fleet management)

This commit ensures that google-startup-scripts.service completes before
the services listed above start. Without this change, it is relatively
common for startup-scripts that install packages with yum or apt to fail
due to a conflict over package database lock.

The existing block on apt-daily.service is probably sufficient since
apt-daily-upgrade.service (which actually writes to the package
database) runs After (in SystemD sense) it. But it's harmless to list it
and it's the "real" block. It is also harmless to list units that are
inactive or not defined (e.g. for the wrong operating system)