sshTrustedCA: add pipe event handler (#222)
The integration between sshd and metadata uses a named pipe, that
gives us the flexibility to read the certificate data from metadata on
demand.

The implementation lays on top of the events layer, when a open
operation is performed on the pipe the sshTrustedCA watcher will
open as write only, notify its subscribers of the event(passing down
the open file), the subscriber is responsible to close the file. If
multiple subscribers exist their implementation should coordinate when
and how to close the pipe.