metadata event handler: updates and bug fixes
Metadata handler:
We shouldn't do anything (i.e. updating configuration) if the metadata
passed in by the event watcher is nil.

Event Manager:
Fix an issue with event manager layer where the context and the sync
channels were being handled as mutually exclusive causing the event
manager to never "abort" after a cancel if the callbacks handler go
routines are already setup and running.

Event Handler Test:
Added a test to exercise a call to cancel the context after the
callback handlers go routines are already setup and running.