metadata: remove the sleep around metadata in instance setup (#232)
We introduced a retry strategy within metadata client before, we don't
need to implement retries in the core implementation.

This patch removes the sleep in the `agentInit()` function, we still
keep retrying if we can't get a metadata descriptor for backward
compatibility - it will get changed whenever we change the network
initialization code, for now we are just avoiding the `sleep()` call.

Additionally this patch removes the unwanted and not necessary method
in the `requestConfig` struct in the metadata package since we are
using this infrastructure code for "GET" requests.