metadata: don't try to use metadata before agentInit() is done
Avoid adding metadata dependent code before agentInit() is done and
the routes are not certain to be ready. This avoid blocking too long
before sending systemd-notify signal.

The logger then requires to have another attempt to configure the
cloud logging client after we have metadata ready/available since the
project name is a dependency to it and we use metadata to determine
its value/name.