Events interface improvement (#290)
* events: decouple the watcher execution code from the Run() function

With that now we cann add a AddWatcher()/RemoveWatcher() calls.

* events: introduce new interface calls

It's introducing event package public functions i.e. AddWatcher(),
AddDefaultWatchers() and remove the old initialization functions.

* events: introduce RemoveWatcher()

As a counter part to AddWatcher() this change introduces the
RemoveWatcher() so watchers can be removed in runtime.

* events: add Unsubscribe() call

This far only the subscribed callback could cancel itself, with this
now other areas or management areas of the code can unsubscribe arbitrary
subscribers.

* events: introduce singleton interface

* oslogin: add support to enabling and disabling ca cert

With the new event manager's APIs we can now enable an disable the
sshca watchers and handlers on demand.