instance setup: trust/rely on metadata package's retry (#316)
With this infinite retry we'll fail to handle SIG ABORT. If the
network/MDS is unreachable and a reboot is done guest agent will only
finish after the systemd unit's timeout is reached and then systemd
does a SIG KILL - the MDS package will fail since the context is
canceled and return error, but this loop is ignoring the error and
will keep retrying.

This PR changes the behavior and relies on the metadata package's retry
strategy.