Fix snapshot test failure (#336)
* oslogin: don't attempt to disable/enable if metadata is unreachable

In cases where the network is down during the early initialization stages
we shouldn't attempt to disable/enable oslogin as we don't have the
metadata descriptor available.

* snapshot: simplify listener loop and improve error logging

Simplify the listener control loop and only log error if the error
is not a context cancellation, if this is a context cancellation we
handle it as a normal operation and simply leave the control block.