Network Manager: make sure we clean up ifcfg files (#371)
* Network Manager: make sure we clean up ifcfg files

Previously we were installing ifcfg files to secondary nics to (between
other things) prevent NetworkManager to manage the connection, this change
guarantees that we remove these files to allow NetworkManager to manage
the connection. This makes sure the implementation works for package/image
update paths.

* NetworkManager: use ifname instead of ip when setting it to UP

It seems the id we are using is not being honored when doing "nmcli
conn up", lets use the interface name (ifname) instead.

Failing to set the connection up also made the ip forwarding setup
to never be executed/performed.