network: introduce multi manager support
We have a use case of systemd-networkd + netplan managed configuration.
This change allows us to chain more than one network manager when
required. The rollback strategy has changed as well to apply the
following semantics: select all managers that apply to the running
system and rollback all the others that don't apply.

The rollback procedures were adjusted to avoid failing whenever a
configuration is not present.