Network manager netplan implementation (#386)
* network: introduce multi manager support

We have a use case of systemd-networkd + netplan managed configuration.
This change allows us to chain more than one network manager when
required. The rollback strategy has changed as well to apply the
following semantics: select all managers that apply to the running
system and rollback all the others that don't apply.

The rollback procedures were adjusted to avoid failing whenever a
configuration is not present.

* network: adjust Rollback() implementations

Considering now that the logic of Rollback() calls has changed to be
called to all the non active network managers the interface
implementation is changing to not report error if the configuration
is not found - or the pre condition is not met.

* systemd-networkd: set priority higher than netplans

For systems running systemd-networkd alongside with netplan will
have their configuration managed by netplan instead, we keep
installing the .network and .netdev files as netplan can be removed.

* netplan: introduce initial implementation

* systemd-networkd: rollback previously supported priority/files.

* network: revert the managed by guest agent section to original format

Since we've had already released a version with the old section
definition we should keep honoring it.

* systemd-networkd: unify the path vs filepath use.