Network subsystem remove os rules (#396)
* netplan: add support for default MTU value from mds

MDS will provide us with the user's defined MTU make sure we honor
it on the netplan dropin.

* network: remove OS rules and introduce netplan dhcp overrides

We are removing the backward compatibility OS rules and assuming the
detected available backend. This change also removes the notion of
fallback as it was before. Additionally the rollback logic changed
to "rollback all backends except the active one", the "except the
active one" rule is important to prevent restarting network when
no configuration change's happened.

In the set we are also adding the netplan dhcp overrides directives
that was resulting into misconfigured dhcp routes.