network: restore debian-12 netplan configuration. (#433)
In a previous release guest-agent team introduced a code path that
removed the debian-12's netplan default configuration instead of
overriding it.

This change makes sure the default configuration is re-created and
bring it consistent to the default experience of debian-12 images.

Additionally a new configuration key was introduced to allow users
to prevent guest-agent creating the configuration:
   NetworkInterfaces.restore_debian12_netplan_config

The new configuration is set to true by default.