network: only trigger networkctl reload for specific interface
Instead of blindly reloading all network configurations with networkctl
reload we now provide the interface name we want to have reload - that
avoids disturbing network configurations not touched our managed by
guest-agent.

For now we are accounting only for the regular interface configuration,
vLAN interfaces and the rollback code paths are intentionally unchanged.