networkd(vlan): remove the interface in addition to config
By only removing the configuration file and reloading it will not
remove the actual interface, for that we need to ask networkd to
actively remove the interface for us.

In order to maintain consistency of what's presend on disk and the
MDS representation + guest agent state we are forcing the removal
of the interface first and then removing the config, in case of
error during networkctl call we keep using the on disk state to
determine what to remove/keep.