network: force address manager to always consolidate the OS state (#585)
Previously we would only run the manager in case the MDS descriptors
changed, the problem with this approach is that it would not consider
changes in the underlying OS.

Always returning Diff() in the manager API we force the manager to
always run and consolidate the expected routes/address configuration
with the expected state described in the MDS keys. With that even
if a route is removed the consolidation process will re add them
every MDS longpoll cycle (every 60 seconds).